feat: 重构 AI 角色聊天与认证体系

+++ src/main/java/com/xzh/friendxxx/service/SocialPostService.java
@@ -26,7 +26,7 @@ public interface SocialPostService extends IService<SocialPost> {
 
     void updateAvatarUrl(String avatarUrl);
 
-    SocialPost getByCurrentId(Integer id);
+    SocialPost getByCurrentId(Long id);
 
     int updateLikescount(LikesDTO likesDTO);
